various reorganizations, renames, and improvements
