+++ tests/test_buffered_text_reader.py
@@ -35,9 +35,9 @@ def single_line_text_file_path(tmp_path_factory: pytest.TempPathFactory):
 
 
 def test_buffered_text_reader(large_text_file_path: pathlib.Path):
-    maximum_ram_usage_in_bytes = 10**6  # 1 MB
+    maximum_buffer_size_in_bytes = 10**6  # 1 MB
     buffered_text_reader = dandi_s3_log_parser.BufferedTextReader(
-        file_path=large_text_file_path, maximum_ram_usage_in_bytes=maximum_ram_usage_in_bytes
+        file_path=large_text_file_path, maximum_buffer_size_in_bytes=maximum_buffer_size_in_bytes
     )
 
     assert iter(buffered_text_reader) is buffered_text_reader, "BufferedTextReader object is not iterable!"
@@ -55,15 +55,15 @@ def test_buffered_text_reader(large_text_file_path: pathlib.Path):
 
 
 def test_value_error(single_line_text_file_path: pathlib.Path):
-    maximum_ram_usage_in_bytes = 10**6  # 1 MB
+    maximum_buffer_size_in_bytes = 10**6  # 1 MB
     with pytest.raises(ValueError) as error_info:
         buffered_text_reader = dandi_s3_log_parser.BufferedTextReader(
-            file_path=single_line_text_file_path, maximum_ram_usage_in_bytes=maximum_ram_usage_in_bytes
+            file_path=single_line_text_file_path, maximum_buffer_size_in_bytes=maximum_buffer_size_in_bytes
         )
         next(buffered_text_reader)
 
     expected_message = (
         "BufferedTextReader encountered a line at offset 0 that exceeds the buffer size! "
-        "Try increasing the `buffer_size_in_bytes` to account for this line."
+        "Try increasing the `maximum_buffer_size_in_bytes` to account for this line."
     )
     assert str(error_info.value) == expected_message
